+++ src/Compilers/Metadata/CustomDebugInfoReader.cs
@@ -823,56 +823,4 @@ namespace Microsoft.CodeAnalysis.Debugging
             return string.Format("0x{0:x8}", methodToken);
         }
     }
-
-    /// <remarks>
-    /// Exposed for <see cref="T:Roslyn.Test.PdbUtilities.PdbToXmlConverter"/>.
-    /// </remarks>
-    internal struct CustomDebugInfoRecord
-    {
-        public readonly CustomDebugInfoKind Kind;
-        public readonly byte Version;
-        public readonly ImmutableArray<byte> Data;
-
-        public CustomDebugInfoRecord(CustomDebugInfoKind kind, byte version, ImmutableArray<byte> data)
-        {
-            this.Kind = kind;
-            this.Version = version;
-            this.Data = data;
-        }
-    }
-
-    internal enum VBImportScopeKind
-    {
-        Unspecified,
-        File,
-        Project,
-    }
-
-    internal struct StateMachineHoistedLocalScope
-    {
-        public readonly int StartOffset;
-        public readonly int EndOffset;
-
-        public StateMachineHoistedLocalScope(int startoffset, int endOffset)
-        {
-            this.StartOffset = startoffset;
-            this.EndOffset = endOffset;
-        }
-    }
-
-    internal struct DynamicLocalInfo
-    {
-        public readonly int FlagCount;
-        public readonly ulong Flags;
-        public readonly int SlotId;
-        public readonly string Name;
-
-        public DynamicLocalInfo(int flagCount, ulong flags, int slotId, string name)
-        {
-            this.FlagCount = flagCount;
-            this.Flags = flags;
-            this.SlotId = slotId;
-            this.Name = name;
-        }
-    }
 }
