+++ src/Compilers/Metadata/VBImportScopeKind.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+namespace Microsoft.CodeAnalysis.Debugging
+{
+    internal enum VBImportScopeKind
+    {
+        Unspecified,
+        File,
+        Project,
+    }
+}
